feat: move kit-modal components into kit-ui
