+++ .changeset/tall-bears-hear.md
@@ -1,0 +1,6 @@
+---
+'@rainbow-me/kit-core': patch
+'@rainbow-me/kit-ui': patch
+---
+
+Move kit-modal components into kit-ui
