+++ packages/core/package.json
@@ -31,7 +31,6 @@
     "@ethersproject/providers": "^5.5.0",
     "@ethersproject/units": "^5.5.0",
     "@rainbow-me/kit-hooks": "workspace:*",
-    "@rainbow-me/kit-modal": "workspace:*",
     "@rainbow-me/kit-ui": "workspace:*",
     "@rainbow-me/kit-utils": "workspace:*",
     "@web3-react/abstract-connector": "^6.0.7"
