+++ packages/core/src/index.ts
@@ -1,5 +1,3 @@
 export * from '@rainbow-me/kit-utils'
 export * from '@rainbow-me/kit-hooks'
 export * from '@rainbow-me/kit-ui'
-export { useWalletModal } from '@rainbow-me/kit-modal'
-export type { UseWalletModalOptions } from '@rainbow-me/kit-modal'
