+++ packages/example/package.json
@@ -6,7 +6,6 @@
   "main": "index.js",
   "dependencies": {
     "@rainbow-me/kit-core": "workspace:*",
-    "@rainbow-me/kit-modal": "workspace:*",
     "@rainbow-me/kit-ui": "workspace:*",
     "@rainbow-me/kit-utils": "workspace:*",
     "@vanilla-extract/next-plugin": "^1.0.1",
