+++ packages/example/pages/_app.tsx
@@ -1,4 +1,3 @@
-import '@rainbow-me/kit-modal/index.css'
 import '@rainbow-me/kit-ui/index.css'
 import type { AppProps } from 'next/app'
 import React from 'react'
