--- packages/modal/CHANGELOG.md
@@ -1,180 +1,0 @@
-# @rainbow-me/kit-modal
-
-## 0.0.25
-
-### Patch Changes
-
-- fix walletconnect in modal
-- Updated dependencies
-  - @rainbow-me/kit-utils@0.0.14
-
-## 0.0.24
-
-### Patch Changes
-
-- bump vanilla-extract
-
-## 0.0.23
-
-### Patch Changes
-
-- ultimatively fix esbuild
-- Updated dependencies
-  - @rainbow-me/kit-utils@0.0.13
-
-## 0.0.22
-
-### Patch Changes
-
-- Updated dependencies
-  - @rainbow-me/kit-utils@0.0.12
-
-## 0.0.21
-
-### Patch Changes
-
-- chore: bump deps and fix failing utils build
-- Updated dependencies
-  - @rainbow-me/kit-utils@0.0.11
-
-## 0.0.20
-
-### Patch Changes
-
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.10
-
-## 0.0.19
-
-### Patch Changes
-
-- fix(modal): pseudoselector
-
-## 0.0.18
-
-### Patch Changes
-
-- fix type import in modal
-
-## 0.0.17
-
-### Patch Changes
-
-- vanilla-extract and many other updates
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.9
-
-## 0.0.16
-
-### Patch Changes
-
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.8
-
-## 0.0.15
-
-### Patch Changes
-
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.7
-
-## 0.0.14
-
-### Patch Changes
-
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.6
-
-## 0.0.13
-
-### Patch Changes
-
-- fix minor bugs and add basic tests
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.5
-
-## 0.0.12
-
-### Patch Changes
-
-- fix: types and bad ifs in strict mode
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.4
-
-## 0.0.11
-
-### Patch Changes
-
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.3
-
-## 0.0.10
-
-### Patch Changes
-
-- feat(ui): make Profile look like in the mockup
-
-## 0.0.9
-
-### Patch Changes
-
-- make EmojiIcon more useful and remove Torus from supported connectors of web3-react (for now)
-
-## 0.0.8
-
-### Patch Changes
-
-- feat(ui, modal): decouple components from Profile, improve default styles for Modal
-
-## 0.0.7
-
-### Patch Changes
-
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.2
-
-## 0.0.6
-
-### Patch Changes
-
-- fix(modal): don't show "0" if there are no hidden wallets
-
-## 0.0.5
-
-### Patch Changes
-
-- fix not connecting on mount for modal hook
-
-## 0.0.4
-
-### Patch Changes
-
-- make the user import connectors instead of a library doing it
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.1
-
-## 0.0.3
-
-### Patch Changes
-
-- do not try to import if module is undefined
-
-## 0.0.2
-
-### Patch Changes
-
-- bug release
-
-## 0.0.1
-
-### Patch Changes
-
-- try to prefix import path of importConnector with `./node_modules`
-
-## 0.0.0
-
-### Patch Changes
-
-- initial pre-release
-- Updated dependencies [undefined]
-  - @rainbow-me/kit-utils@0.0.0
