--- packages/modal/assets.d.ts
@@ -1,15 +1,0 @@
-declare global {
-  interface Window {
-    ethereum: ExternalProvider
-  }
-}
-
-declare module '*.svg' {
-  const content: string
-  export default content
-}
-
-declare module '*.png' {
-  const content: string
-  export default content
-}
