--- packages/modal/package.json
@@ -1,61 +1,0 @@
-{
-  "name": "@rainbow-me/kit-modal",
-  "version": "0.0.25",
-  "description": "Universal Ethereum wallet modal library for React",
-  "files": [
-    "dist"
-  ],
-  "type": "module",
-  "exports": {
-    ".": "./dist/index.js",
-    "./index.css": "./dist/index.css"
-  },
-  "types": "./dist/index.d.ts",
-  "engines": {
-    "node": ">=12.4"
-  },
-  "scripts": {
-    "build": "node build.js",
-    "build:watch": "node build.js --watch",
-    "dev": "pnpm build:watch & pnpm typegen:watch",
-    "prepare": "pnpm build",
-    "prebuild": "pnpm typegen",
-    "typecheck": "pnpm tsc --noEmit",
-    "typegen": "tsc --emitDeclarationOnly || true",
-    "typegen:watch": "tsc --emitDeclarationOnly --watch"
-  },
-  "keywords": [
-    "ens",
-    "ethereum",
-    "react",
-    "react-hook",
-    "hooks",
-    "blockchain"
-  ],
-  "author": "v1rtl",
-  "license": "UNLICENSED",
-  "peerDependencies": {
-    "react": ">=17"
-  },
-  "devDependencies": {
-    "react": "^17.0.2"
-  },
-  "dependencies": {
-    "@rainbow-me/kit-utils": "workspace:*",
-    "@vanilla-extract/css": "^1.6.6",
-    "@vanilla-extract/dynamic": "^2.0.1",
-    "@web3-react/abstract-connector": "^6.0.7",
-    "clsx": "^1.1.1"
-  },
-  "optionalDependencies": {
-    "@web3-react/frame-connector": "^6.0.9",
-    "@web3-react/injected-connector": "^6.0.7",
-    "@web3-react/walletconnect-connector": "6.2.4",
-    "@web3-react/walletlink-connector": "^6.2.8"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/rainbow-me/rainbowkit.git",
-    "directory": "packages/modal"
-  }
-}
