--- packages/modal/src/index.ts
@@ -1,3 +1,0 @@
-export * from './useWalletModal'
-export * from './components/Modal'
-export * from './types'
