--- packages/modal/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "module": "ESNext",
-    "target": "ES2019",
-    "declaration": true,
-    "declarationDir": "dist",
-    "jsx": "react",
-    "outDir": "dist"
-  },
-  "exclude": ["node_modules"],
-  "include": ["src", "assets.d.ts"]
-}
