+++ packages/ui/package.json
@@ -35,23 +35,27 @@
   "author": "v1rtl",
   "license": "UNLICENSED",
   "peerDependencies": {
+    "@web3-react/core": ">=6.1",
     "react": ">=17"
   },
   "devDependencies": {
     "@vanilla-extract/private": "^1.0.2",
+    "@web3-react/core": "^6.1.9",
     "autoprefixer": "^10.4.0",
     "postcss": "^8.4.4",
     "react": "^17.0.2"
   },
   "dependencies": {
     "@ethersproject/bignumber": "^5.5.0",
+    "@ethersproject/providers": "^5.5.0",
     "@ethersproject/transactions": "^5.5.0",
     "@ethersproject/units": "^5.5.0",
     "@rainbow-me/kit-hooks": "workspace:*",
-    "@rainbow-me/kit-modal": "workspace:*",
     "@rainbow-me/kit-utils": "workspace:*",
     "@vanilla-extract/css": "^1.6.6",
     "@vanilla-extract/sprinkles": "^1.3.2",
+    "@web3-react/injected-connector": "^6.0.7",
+    "@web3-react/walletconnect-connector": "^6.2.4",
     "clsx": "^1.1.1",
     "focus-visible": "^5.2.0",
     "get-ens": "^2.0.3"
