+++ packages/ui/src/components/Modal/index.ts
@@ -1,0 +1,2 @@
+export * from './useWalletModal'
+export * from './Modal'
