+++ packages/ui/src/index.ts
@@ -6,8 +6,7 @@ export * from './components/CopyAddressButton'
 export * from './components/TxHistory'
 export * from './components/NetworkSelect'
 export * from './components/Profile'
-export { Modal } from '@rainbow-me/kit-modal'
-export type { ModalProps } from '@rainbow-me/kit-modal'
 export * from './components/WalletDropdown'
 export * from './components/Badge'
 export * from './components/EmojiIcon'
+export * from './components/Modal'
